About Component and Dark Mode Feature
Added a About component to our React app as well as created a button on the about page which helps to toggle between the dark and light mode of the about page. In future, will bring it to the whole app, but currently this feature only available to the about page.